deforestation_segmentation/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.11.2
jupyter>=1.0.0

# Image processing
rasterio>=1.2.6
geopandas>=0.9.0
pyproj>=3.1.0
albumentations>=1.3.0
Pillow>=9.5.0

# Deep learning
tensorboard>=2.13.0
tqdm>=4.65.0
pyyaml>=5.4.1

# Testing and development
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0

# Documentation
sphinx>=4.1.2
sphinx-rtd-theme>=0.5.2

# Optional: GPU support
# cuda-python>=11.4.0  # Uncomment if using CUDA 

# Added from the code block
timm>=0.9.0
opencv-python>=4.7.0
wandb>=0.15.0

# Web application
streamlit>=1.24.0
plotly>=5.15.0
opencv-python-headless>=4.7.0
geopy>=2.3.0
geopandas>=0.13.0
rasterio>=1.3.0
folium>=0.14.0

# Segmentation specific
segmentation-models-pytorch>=0.3.0
pytorch-lightning>=2.0.0
monai>=1.2.0  # Additional medical imaging tools useful for segmentation 